Prepared to deploy slides to the internal directory
When a slides file (*.ppt or *.pptx) is encountered, it is recognized as distinct from other files. It will be deployed to the directory with Live Coder's special files and opened from that location. In that way, slides will be deployed aside from source code, so that they do not accidentaly appear in the project tree.