Deployer loads XML snippets from the *.snippet file
The *.snippet file is an XML file which can contain many separate snippets. Each snippet is identified by a shortcut in format "snpNN", and code. Shortcut is parsed into an int, and code is extracted as a stirng. These two elements will later be used to generate the demo script.