Renamed Deployment to DeploymentSpecification
By adding the term "Specification", we are closely explaining that object of this type is only specifying the deployment, but not representing the result of deploying the demo. This specification is executable, i.e. it exposes the Execute() method, which will result in building another object that represents the deployed demo.