Deployer is redeploynig XML snippets when changed
When XML snippets are translated to script and deployed, the originating *.snippet file is tracked for changes. When the file is modified, it is translated and deployed again. This process may inculde clashes on file access, as the *.snippet file may still be temporarily locked. Reading the file is postponed until the lock is removed, and only then the deployment is performed.