Snippets deployment skipped if final file not modified
When snippets are translated to script, the resulting script is compared to content of the script file. If script file doesn't exist, or its content is different than what is about to be written, only then the new content is written to the script file. This measure is meant to help the VS extension avoid multiple refreshes of the script content in cases when there is no actual modification to the content.