Implemented snippets normalization
XML snippets are supposed to be numbered snp01, snp02, snp03, ... Sometimes, during snippets development, author wishes to insert a snippet into the middle of the sequence, e.g. between snippets snp01 and snp02. Such snippets should then be inserted between snp01 and snp02 in the snippets XML, and numbered snp01a, snp01b, snp01c, etc. so that the lexicographic sort order is preserved. The deployer will detect those inserted shortcuts and renumerate all snippets so that the correct sequence is established. It will then overwrite the snippets file, so that the author can continue to edit them in usual way.