Deployment directory name format changed
Deployer is creating new output directory which is named by the timestamp when it was created. Earlier format was the timestamp with 1-second precision, while termporary format during the new deployer development was 1-millisecond precision timestamp and ".new" added to the directory name. These additions are now removed, and output directory will again be named like: C:\Demo\yyyyMMddHHmmss