Allowed multiple declarations of the same snippet
Demo script will overwrite any snippet when the new snippet with the same number is added to it. This feature is a preparation for the further development of the script interpreter. The plan is to declare snippets in two ways: existing declaration (the snippet command) and the new way using method-like calls to add snippets to the collection. Once both ways are in place, the old one will be retired.