Parser: Missing dot means next token applies to global scope
When parsing an array of tokens, parser is observing the next coming token. If it is a dot, then whatever comes after it will apply to current scope. Otherwise, it applies to global scope. E.g. code obj.attrib first applies to global scope, where 'obj' is searched; then the rest ('.attrib') begins with a dot and applies to local scope (which is the result of evaluating 'obj').