Parser: Introduced SLR(1) parser
Parser is capable of producing a complete syntax tree for cases that are already covered. This tree will be converted to script instructions later. When that happens, this SLR(1) parser will become official, and its output will be used in script execution. Currently, generated syntax tree is not used, but simply discarded.