Migrated code snippets engine to the Scripting project
All classes and interfaces used to implement code snippets have been moved to the Scripting project. This project is the base for all types that are executing demo scripts. The VS Extension project is planned to remain a thin implementation of the interfaces that gain access to the source code upon which the script is executed.