Solution is not exposing solution file location
Solution file was used to identify the .livecoder directory relative to its location. This duty is now passed on to the VS extension, which should know where the script and accompanying files are located (including the directory name). This directory is passed to demo engines on instantiation, so that they can load their components.