Reminders can be put on lines that contain code
Reminder snippet (e.g. snp07.3) do not substitute code with a snippet from the script file. Instead, their only purpose is to place cursor on a specific line, which should be discussed. New change allows that line to contain code. When reminder is executed, the line will not be deleted, but only the snippet reference (i.e. the comment) will be removed from it. This makes reminders more compact than they used to be.