Skipping verification after it has failed once
I haven't found the cause for intermittent verification step failures (e.g. document which is active claims that it is not). To work around possible show-stoppers in live presentations, only the first verification failure will cause current commands to be purged. Subsequent failure on the same step will be ignored, and step will be executed nevertheless.